# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GST Invoice System API",
    "version": "1.0.0",
...
    "license": {
      "name": "See LICENSE in the repository"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://{host}",
      "description": "Your deployment",
      "variables": {
        "host": {
          "default": "invoices.example.com"
        }
      }
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Set by POST /api/auth/login. HttpOnly, SameSite, and revocable server-side."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message, safe to show a user."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, present on errors a client is expected to handle (PLAN_LIMIT, PERIOD_LOCKED, VALUATION_RESTATEMENT, RETENTION_WARNING, DUPLICATE, HAS_MOVEMENTS, ALREADY_POSTED)."
          }
        }
      },
      "ValidationError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Error"
          },
          {
            "type": "object",
            "properties": {
              "issues": {
                "type": "array",
                "description": "Zod issues, when the failure was schema validation.",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "Quantity": {
        "type": "number",
        "description": "A stock quantity. Fractional, because real units are (2.5 kg, 0.75 m). Not money - quantities are the only numbers in this API that are not integers.",
        "example": 2.5
      },
...
      "DocumentLine": {
        "type": "object",
        "required": [
          "itemName",
          "quantity",
          "rate"
        ],
        "properties": {
          "itemId": {
            "type": "string",
            "nullable": true,
            "description": "Link to an Item to move stock. Omit for a service line."
          },
          "itemName": {
            "type": "string"
          },
          "hsn": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "$ref": "#/components/schemas/Quantity"
          },
          "rate": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "description": "Unit rate in RUPEES as typed by the user. Strings are parsed digit-wise so \"0.145\" is exact. Converted to paise server-side; the response is paise."
          },
          "discount": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "description": "Line discount in rupees. Applied BEFORE tax, per section 15(3) of the CGST Act."
          },
          "gstRate": {
            "type": "number"
          },
          "cessRate": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "supplyType": {
            "type": "string",
            "enum": [
              "TAXABLE",
              "EXEMPT",
              "NIL_RATED",
              "NON_GST",
              "ZERO_RATED"
            ]
          },
          "pricingMode": {
            "type": "string",
            "enum": [
              "EXCLUSIVE",
              "INCLUSIVE"
            ]
          }
        }
      },
...
      "CreateInvoice": {
        "type": "object",
        "required": [
          "partyId",
          "lines"
        ],
        "properties": {
          "partyId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/DocumentLine"
            }
          },
          "discount": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "description": "Invoice-level discount in rupees, apportioned across lines before tax."
          },
          "additionalCharges": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ]
          },
          "tdsRate": {
            "type": "number"
          },
          "reverseCharge": {
            "type": "boolean"
          },
          "placeOfSupply": {
            "type": "string",
            "description": "Two-digit state code. Defaults to the party's."
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/invoices/{id}/reminder": {
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Send a payment reminder",
        "description": "Rate limited per company and per IP: the outbound address is shared across tenants.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource id. Always scoped to the caller's company."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or belongs to another company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/credit-notes": {
      "get": {
        "tags": [
          "Sales"
        ],
        "summary": "List credit and debit notes",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "1-based page number."
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "description": "Rows per page. Clamped to 200."
          },
          {
            "name": "kind",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CREDIT",
                "DEBIT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "page",
                    "pageSize",
                    "total",
                    "totalPages",
                    "hasMore"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    },
                    "totalPages": {
                      "type": "integer"
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sales"
        ],
        "summary": "Raise a credit or debit note",
        "description": "CREDIT is a sales return: goods come back in, at COST rather than at the credited price. DEBIT is a purchase return.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateInvoice"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "kind": {
                        "type": "string",
                        "enum": [
                          "CREDIT",
                          "DEBIT"
                        ]
                      },
                      "reason": {
                        "type": "string"
                      },
                      "originalRef": {
                        "type": "string"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Read-only role, or the company is suspended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Export a dataset as CSV",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "invoices",
                "parties",
                "items",
                "purchases",
                "payments",
                "expenses"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the session was revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}